Implemented basic raytracing

+++ raytracer/core/ICanvas.cs
@@ -4,7 +4,7 @@ namespace Atamocius.Core
 {
     public interface ICanvas
     {
-        (uint Width, uint Height) Size { get; }
+        (ushort Width, ushort Height) Size { get; }
         void Clear(in Color color);
         void Present(RenderTarget ctx);
         void PutPixel(in int x, in int y, in Color color);
